+++ pkg/controller/controller_test.go
@@ -1,19 +1,15 @@
 package controller
 
 import (
-	"fmt"
-	"istio.io/istio/pkg/config/schema/collections"
 	"testing"
 	"time"
 
-	"istio.io/api/rbac/v1alpha1"
 	"istio.io/istio/pilot/pkg/model"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
 
-	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac/common"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/log"
 	adv1 "github.com/yahoo/k8s-athenz-syncer/pkg/apis/athenz/v1"
 	"github.com/yahoo/k8s-athenz-syncer/pkg/client/clientset/versioned/fake"
@@ -66,299 +62,6 @@ func TestProcessConfigEvent(t *testing.T) {
 	assert.False(t, shutdown, "shutdown should be false")
 	assert.Equal(t, 0, c.queue.Len(), "queue length should be 0")
 	assert.Equal(t, "test.namespace", item, "key should be equal")
-}
-
-func newSr(ns, role string) model.Config {
-	srSpec := &v1alpha1.ServiceRole{
-		Rules: []*v1alpha1.AccessRule{
-			{
-				Services: []string{common.WildCardAll},
-				Methods:  []string{"GET"},
-				Constraints: []*v1alpha1.AccessRule_Constraint{
-					{
-						Key:    common.ConstraintSvcKey,
-						Values: []string{"test-svc"},
-					},
-				},
-			},
-		},
-	}
-	return common.NewConfig(collections.IstioRbacV1Alpha1Serviceroles, ns, role, srSpec)
-}
-
-func newSrb(ns, role string) model.Config {
-	srbSpec := &v1alpha1.ServiceRoleBinding{
-		RoleRef: &v1alpha1.RoleRef{
-			Kind: common.ServiceRoleKind,
-			Name: role,
-		},
-		Subjects: []*v1alpha1.Subject{
-			{
-				User: "test-user",
-			},
-		},
-	}
-	return common.NewConfig(collections.IstioRbacV1Alpha1Servicerolebindings, ns, role, srbSpec)
-}
-
-func updatedSr(ns, role string) model.Config {
-	srSpec := &v1alpha1.ServiceRole{
-		Rules: []*v1alpha1.AccessRule{
-			{
-				Services: []string{common.WildCardAll},
-				Methods:  []string{"GET"},
-				Constraints: []*v1alpha1.AccessRule_Constraint{
-					{
-						Key:    common.ConstraintSvcKey,
-						Values: []string{"test-svc"},
-					},
-				},
-			},
-			{
-				Services: []string{common.WildCardAll},
-				Methods:  []string{"POST"},
-				Constraints: []*v1alpha1.AccessRule_Constraint{
-					{
-						Key:    common.ConstraintSvcKey,
-						Values: []string{"test-svc"},
-					},
-				},
-			},
-		},
-	}
-	return common.NewConfig(collections.IstioRbacV1Alpha1Serviceroles, ns, role, srSpec)
-}
-
-func updatedSrb(ns, role string) model.Config {
-	srbSpec := &v1alpha1.ServiceRoleBinding{
-		RoleRef: &v1alpha1.RoleRef{
-			Kind: common.ServiceRoleKind,
-			Name: role,
-		},
-		Subjects: []*v1alpha1.Subject{
-			{
-				User: "test-user",
-			},
-			{
-				User: "another.client.user",
-			},
-		},
-	}
-	return common.NewConfig(collections.IstioRbacV1Alpha1Servicerolebindings, ns, role, srbSpec)
-}
-
-func TestEqual(t *testing.T) {
-	tests := []struct {
-		name     string
-		in1      model.Config
-		in2      model.Config
-		expected bool
-	}{
-		{
-			name:     "should return true for empty model.Config items",
-			in1:      model.Config{},
-			in2:      model.Config{},
-			expected: true,
-		},
-		{
-			name:     "should return false for different model.Config item names but same spec",
-			in1:      newSr("test-ns", "my-role"),
-			in2:      newSr("test-ns", "another-role"),
-			expected: false,
-		},
-		{
-			name:     "should return false for different model.Config item names and different spec",
-			in1:      newSr("test-ns", "my-role"),
-			in2:      updatedSr("test-ns", "another-role"),
-			expected: false,
-		},
-		{
-			name:     "should return false for same model.Config item names but different spec",
-			in1:      newSr("test-ns", "my-role"),
-			in2:      updatedSr("test-ns", "my-role"),
-			expected: false,
-		},
-		{
-			name:     "should return false for different model.Config item types but same names",
-			in1:      newSr("test-ns", "my-role"),
-			in2:      newSrb("test-ns", "my-role"),
-			expected: false,
-		},
-		{
-			name:     "should return true for same model.Config item names and spec",
-			in1:      newSr("test-ns", "my-role"),
-			in2:      newSr("test-ns", "my-role"),
-			expected: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			actual := common.Equal(tt.in1, tt.in2)
-			assert.Equal(t, tt.expected, actual, "comparison result should be equal to expected")
-		})
-	}
-}
-
-func TestComputeChangeList(t *testing.T) {
-
-	cbHandler := func(err error, item *common.Item) error {
-		return err
-	}
-
-	type input struct {
-		current   []model.Config
-		desired   []model.Config
-		cbHandler common.OnCompleteFunc
-	}
-	tests := []struct {
-		name           string
-		input          input
-		expectedOutput []*common.Item
-	}{
-		{
-			name:           "should return empty change list for empty current and desired list",
-			input:          input{},
-			expectedOutput: make([]*common.Item, 0),
-		},
-		{
-			name: "should add create operations for new items on the desired list",
-			input: input{
-				current: []model.Config{},
-				desired: []model.Config{
-					newSr("test-ns", "svc-role"),
-					newSrb("test-ns", "svc-role"),
-				},
-				cbHandler: cbHandler,
-			},
-			expectedOutput: []*common.Item{
-				{
-					Operation:       model.EventAdd,
-					Resource:        newSr("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventAdd,
-					Resource:        newSrb("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-			},
-		},
-		{
-			name: "should add update operations for changed items on the desired list",
-			input: input{
-				current: []model.Config{
-					newSr("test-ns", "svc-role"),
-					newSrb("test-ns", "svc-role"),
-					newSr("another-ns", "backend-writer"),
-					newSrb("another-ns", "backend-writer"),
-				},
-				desired: []model.Config{
-					updatedSr("test-ns", "svc-role"),
-					updatedSrb("test-ns", "svc-role"),
-					newSr("another-ns", "backend-writer"),
-					newSrb("another-ns", "backend-writer"),
-				},
-				cbHandler: cbHandler,
-			},
-			expectedOutput: []*common.Item{
-				{
-					Operation:       model.EventUpdate,
-					Resource:        updatedSr("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventUpdate,
-					Resource:        updatedSrb("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-			},
-		},
-		{
-			name: "should add delete operation for deleted items on the desired list",
-			input: input{
-				current: []model.Config{
-					newSr("test-ns", "svc-role"),
-					newSrb("test-ns", "svc-role"),
-				},
-				desired:   []model.Config{},
-				cbHandler: cbHandler,
-			},
-			expectedOutput: []*common.Item{
-				{
-					Operation:       model.EventDelete,
-					Resource:        newSr("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventDelete,
-					Resource:        newSrb("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-			},
-		},
-		{
-			name: "should add create,update and delete operations based on the desired list",
-			input: input{
-				current: []model.Config{
-					newSr("test-ns", "svc-role"),
-					newSrb("test-ns", "svc-role"),
-					newSr("some-ns", "frontend-reader"),
-					newSrb("some-ns", "frontend-reader"),
-				},
-				desired: []model.Config{
-					updatedSr("test-ns", "svc-role"),
-					updatedSrb("test-ns", "svc-role"),
-					newSr("another-ns", "backend-writer"),
-					newSrb("another-ns", "backend-writer"),
-				},
-				cbHandler: cbHandler,
-			},
-			expectedOutput: []*common.Item{
-				{
-					Operation:       model.EventUpdate,
-					Resource:        updatedSr("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventUpdate,
-					Resource:        updatedSrb("test-ns", "svc-role"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventAdd,
-					Resource:        updatedSr("another-ns", "backend-writer"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventAdd,
-					Resource:        updatedSrb("another-ns", "backend-writer"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventDelete,
-					Resource:        newSr("some-ns", "frontend-reader"),
-					CallbackHandler: cbHandler,
-				},
-				{
-					Operation:       model.EventDelete,
-					Resource:        newSrb("some-ns", "frontend-reader"),
-					CallbackHandler: cbHandler,
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			actualChangeList := common.ComputeChangeList(tt.input.current, tt.input.desired, tt.input.cbHandler, nil)
-			assert.Equal(t, len(tt.expectedOutput), len(actualChangeList), "len(expectedChangeList) and len(actualChangeList) should match")
-			for i, expectedItem := range tt.expectedOutput {
-				assert.Equal(t, expectedItem.Operation, actualChangeList[i].Operation, fmt.Sprintf("operation on changelist[%d] does not match with expected", i))
-				assert.Equal(t, expectedItem.Operation, actualChangeList[i].Operation, fmt.Sprintf("operation on changelist[%d] does not match with expected", i))
-			}
-		})
-	}
 }
 
 func TestResync(t *testing.T) {
