+++ pkg/istio/authorizationpolicy/controller.go
@@ -6,36 +6,29 @@ package authzpolicy
 import (
 	"errors"
 	"fmt"
-	"github.com/ghodss/yaml"
-	"github.com/yahoo/athenz/clients/go/zms"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/athenz"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac/common"
-	rbacv1 "github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac/v1"
+	rbacv2 "github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac/v2"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/log"
 	adv1 "github.com/yahoo/k8s-athenz-syncer/pkg/apis/athenz/v1"
-	"io/ioutil"
 	"istio.io/client-go/pkg/clientset/versioned"
 	istioCache "istio.io/client-go/pkg/informers/externalversions/security/v1beta1"
-	"istio.io/istio/pilot/pkg/config/kube/crd"
 	"istio.io/istio/pilot/pkg/model"
-	"istio.io/istio/pkg/config/schema/collections"
 	corev1 "k8s.io/api/core/v1"
 	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
-	"os"
 	"strings"
 	"time"
 )
 
 const (
-	queueNumRetries            = 3
-	authzEnabled               = "true"
-	authzEnabledAnnotation     = "authz.istio.io/enabled"
-	overrideAnnotation         = "overrideAuthzPolicy"
-	DryRunStoredFilesDirectory = "/root/authzpolicy/"
+	queueNumRetries        = 3
+	authzEnabled           = "true"
+	authzEnabledAnnotation = "authz.istio.io/enabled"
+	overrideAnnotation     = "overrideAuthzPolicy"
 )
 
 type Controller struct {
@@ -47,17 +40,11 @@ type Controller struct {
 	rbacProvider             rbac.Provider
 	apResyncInterval         time.Duration
 	enableOriginJwtSubject   bool
-	dryrun                   bool
+	dryRun                   bool
 }
 
-func NewController(configStoreCache model.ConfigStoreCache, serviceIndexInformer cache.SharedIndexInformer, adIndexInformer cache.SharedIndexInformer, istioClientSet versioned.Interface, apResyncInterval time.Duration, enableOriginJwtSubject bool, DryRun bool) *Controller {
+func NewController(configStoreCache model.ConfigStoreCache, serviceIndexInformer cache.SharedIndexInformer, adIndexInformer cache.SharedIndexInformer, istioClientSet versioned.Interface, apResyncInterval time.Duration, enableOriginJwtSubject bool, dryRun bool) *Controller {
 	queue := workqueue.NewRateLimitingQueue(workqueue.DefaultControllerRateLimiter())
-
-	if DryRun {
-		if _, err := os.Stat(DryRunStoredFilesDirectory); os.IsNotExist(err) {
-			os.MkdirAll(DryRunStoredFilesDirectory, 0644)
-		}
-	}
 
 	authzpolicyIndexInformer := istioCache.NewAuthorizationPolicyInformer(istioClientSet, "", 0, cache.Indexers{})
 
@@ -67,10 +54,10 @@ func NewController(configStoreCache model.ConfigStoreCache, serviceIndexInformer
 		adIndexInformer:          adIndexInformer,
 		authzpolicyIndexInformer: authzpolicyIndexInformer,
 		queue:                    queue,
-		rbacProvider:             rbacv1.NewProvider(enableOriginJwtSubject),
+		rbacProvider:             rbacv2.NewProvider(enableOriginJwtSubject),
 		apResyncInterval:         apResyncInterval,
 		enableOriginJwtSubject:   enableOriginJwtSubject,
-		dryrun:                   DryRun,
+		dryRun:                   dryRun,
 	}
 
 	serviceIndexInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
@@ -97,23 +84,13 @@ func NewController(configStoreCache model.ConfigStoreCache, serviceIndexInformer
 		},
 	})
 
-	authzpolicyIndexInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			c.processEvent(cache.MetaNamespaceKeyFunc, obj)
-		},
-		UpdateFunc: func(_, newObj interface{}) {
-			c.processEvent(cache.MetaNamespaceKeyFunc, newObj)
-		},
-		DeleteFunc: func(obj interface{}) {
-			c.processEvent(cache.DeletionHandlingMetaNamespaceKeyFunc, obj)
-		},
-	})
-
 	return c
 }
 
 func (c *Controller) EventHandler(config model.Config, _ model.Config, e model.Event) {
-	c.queue.Add(config.Key())
+	// authz policy event handler, Key() returns format <type>/<namespace>/<name>
+	// should drop the type and pass <namespace>/<name> only
+	c.queue.Add(strings.Join(strings.Split(config.Key(), "/")[1:], "/"))
 }
 
 // processEvent is responsible for calling the key function and adding the
@@ -172,27 +149,25 @@ func (c *Controller) processNextItem() bool {
 	return true
 }
 
-// draft: remodeled sync function. sync function receives a key string function, key can have three format:
+// sync function receives a key string function, key can have three format:
 // Case 1: for athenzdomain crd, key string is in format: <athenz domain name>, perform a service list scan.
 //         Compute, compare and update authz policy specs based on current state in cluster
-// Case 2: for service resource, key string is in format: <namespace name>/<service name>, look up svc in
-//         cache and generate corresponding authz policy, update based on current state in cluster
-// Case 3: for authorization policy, key string is in format: AuthorizationPolicy/<namespace name>/<service name>,
+// Case 2: for service resource and authorization policy, key string is in format: <namespace name>/<service name>,
 //         look up svc in cache and generate corresponding authz policy, update based on current state in cluster
 func (c *Controller) sync(key string) error {
 	var serviceName, athenzDomainName string
 
 	parseKeyList := strings.Split(key, "/")
 	if len(parseKeyList) > 1 {
-		athenzDomainName = athenz.NamespaceToDomain(parseKeyList[len(parseKeyList)-2])
-		serviceName = parseKeyList[len(parseKeyList)-1]
+		athenzDomainName = athenz.NamespaceToDomain(parseKeyList[0])
+		serviceName = parseKeyList[1]
 	} else {
 		athenzDomainName = key
 	}
 
 	athenzDomainRaw, exists, err := c.adIndexInformer.GetIndexer().GetByKey(athenzDomainName)
 	if err != nil {
-		return err
+		return fmt.Errorf("unable to fetch athenz domain from cache, error: %s", err)
 	}
 
 	if !exists {
@@ -201,7 +176,7 @@ func (c *Controller) sync(key string) error {
 
 	athenzDomain, ok := athenzDomainRaw.(*adv1.AthenzDomain)
 	if !ok {
-		return errors.New("athenz domain cast failed")
+		return errors.New("athenz domain cast failed, domain name: " + athenzDomainName)
 	}
 
 	signedDomain := athenzDomain.Spec.SignedDomain
@@ -219,8 +194,9 @@ func (c *Controller) sync(key string) error {
 		// error checking
 		// add to serviceList
 		svcKeys := c.serviceIndexInformer.GetStore().ListKeys()
+		domainName := athenz.DomainToNamespace(athenzDomainName)
 		for _, svcKey := range svcKeys {
-			if strings.Split(svcKey, "/")[0] == athenz.DomainToNamespace(athenzDomainName) {
+			if strings.Split(svcKey, "/")[0] == domainName {
 				serviceObj, err := c.getSvcObj(svcKey)
 				if err != nil {
 					return fmt.Errorf("error getting service from cache: %s", err.Error())
@@ -233,17 +209,17 @@ func (c *Controller) sync(key string) error {
 	var desiredCRs []model.Config
 	// range over serviceList
 	for _, service := range serviceList {
+		// if svc annotation authz.istio.io/enabled is not present - skip processing and continue
 		if !c.checkAuthzEnabledAnnotation(service) {
 			continue
 		}
-		// if svc annotation authz.istio.io/enabled is not present - skip processing and continue
-		desiredCR := c.rbacProvider.ConvertAthenzModelIntoIstioAuthzPolicy(domainRBAC, service.Namespace, service.Name, service.Labels["svc"])
+		desiredCR := c.rbacProvider.ConvertAthenzModelIntoIstioRbac(domainRBAC, service.Name, service.Labels["svc"])
 		// append to desiredCRs array
-		desiredCRs = append(desiredCRs, desiredCR)
+		desiredCRs = append(desiredCRs, desiredCR...)
 	}
 
 	// get current APs from cache
-	currentCRs := c.rbacProvider.GetCurrentIstioAuthzPolicy(domainRBAC, c.configStoreCache, serviceName)
+	currentCRs := c.rbacProvider.GetCurrentIstioRbac(domainRBAC, c.configStoreCache, serviceName, c.dryRun)
 	cbHandler := c.getCallbackHandler(key)
 	changeList := common.ComputeChangeList(currentCRs, desiredCRs, cbHandler, c.checkOverrideAnnotation)
 
@@ -256,7 +232,7 @@ func (c *Controller) sync(key string) error {
 
 	for _, item := range changeList {
 		log.Infof("Adding resource action to queue: %s on %s for key: %s", item.Operation, item.Resource.Key(), key)
-		c.ProcessConfigChange(item)
+		c.processConfigChange(item)
 	}
 	return nil
 }
@@ -272,34 +248,27 @@ func (c *Controller) checkOverrideAnnotation(existingConfig model.Config) bool {
 }
 
 // ProcessConfigChange receives resource and event action, and perform update on resource
-func (c *Controller) ProcessConfigChange(item *common.Item) error {
+func (c *Controller) processConfigChange(item *common.Item) error {
 	if item == nil {
 		return nil
 	}
 	var err error
-
-	// dry run mode on, create and store files
-	if c.dryrun {
-		switch item.Operation {
-		case model.EventAdd:
-			err = c.createDryrunResource(item.Resource, item.Resource.ConfigMeta.Name, item.Resource.ConfigMeta.Namespace)
-		case model.EventUpdate:
-			err = c.createDryrunResource(item.Resource, item.Resource.ConfigMeta.Name, item.Resource.ConfigMeta.Namespace)
-		case model.EventDelete:
-			err = c.findDeleteDryrunResource(item.Resource.ConfigMeta.Name, item.Resource.ConfigMeta.Namespace)
+	var eHandler common.EventHandler
+	if c.dryRun {
+		eHandler = &common.DryRunHandler{}
+	} else {
+		eHandler = &common.ApiHandler{
+			ConfigStoreCache: c.configStoreCache,
 		}
-		return err
 	}
 
-	// dry run mode off, call apis with operations
 	switch item.Operation {
 	case model.EventAdd:
-		_, err = c.configStoreCache.Create(item.Resource)
+		err = eHandler.Add(item)
 	case model.EventUpdate:
-		_, err = c.configStoreCache.Update(item.Resource)
+		err = eHandler.Update(item)
 	case model.EventDelete:
-		res := item.Resource
-		err = c.configStoreCache.Delete(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), res.Name, res.Namespace)
+		err = eHandler.Delete(item)
 	}
 
 	return err
@@ -364,9 +333,8 @@ func (c *Controller) resync(stopCh <-chan struct{}) {
 		select {
 		case <-t.C:
 			log.Infoln("Running resync for authorization policies...")
-			apListRaw := c.authzpolicyIndexInformer.GetIndexer().List()
-			for _, adRaw := range apListRaw {
-				//c.ProcessConfigChange(model.EventAdd, adRaw)
+			adListRaw := c.adIndexInformer.GetIndexer().List()
+			for _, adRaw := range adListRaw {
 				c.processEvent(cache.MetaNamespaceKeyFunc, adRaw)
 			}
 		case <-stopCh:
@@ -376,7 +344,7 @@ func (c *Controller) resync(stopCh <-chan struct{}) {
 	}
 }
 
-// checkAuthzEnabledAnnotation checks if current servce object has "authz.istio.io/enabled" annotation set
+// checkAuthzEnabledAnnotation checks if current service object has "authz.istio.io/enabled" annotation set
 func (c *Controller) checkAuthzEnabledAnnotation(serviceObj *corev1.Service) bool {
 	if _, ok := serviceObj.Annotations[authzEnabledAnnotation]; ok {
 		if serviceObj.Annotations[authzEnabledAnnotation] == authzEnabled {
@@ -384,55 +352,4 @@ func (c *Controller) checkAuthzEnabledAnnotation(serviceObj *corev1.Service) boo
 		}
 	}
 	return false
-}
-
-func (c *Controller) createDryrunResource(convertedCR model.Config, authzPolicyName string, namespace string) error {
-	convertedObj, err := crd.ConvertConfig(collections.IstioSecurityV1Beta1Authorizationpolicies, convertedCR)
-	if err != nil {
-		return fmt.Errorf("unable to convert authorization policy config to istio objects, resource name: %v", convertedCR.Name)
-	}
-	configInBytes, err := yaml.Marshal(convertedObj)
-	if err != nil {
-		return fmt.Errorf("could not marshal %v: %v", convertedCR.Name, err)
-	}
-	yamlFileName := authzPolicyName + "--" + namespace + ".yaml"
-	return ioutil.WriteFile(DryRunStoredFilesDirectory+yamlFileName, configInBytes, 0644)
-}
-
-func (c *Controller) findDeleteDryrunResource(authzPolicyName string, namespace string) error {
-	yamlFileName := authzPolicyName + "--" + namespace + ".yaml"
-	if _, err := os.Stat(DryRunStoredFilesDirectory + yamlFileName); os.IsNotExist(err) {
-		log.Infof("file %s does not exist in local directory\n", DryRunStoredFilesDirectory+yamlFileName)
-		return nil
-	}
-	log.Infof("deleting file under path: %s\n", DryRunStoredFilesDirectory+yamlFileName)
-	return os.Remove(DryRunStoredFilesDirectory + yamlFileName)
-}
-
-func (c *Controller) deleteAuthzPolicyResource(obj *corev1.Service) error {
-	log.Infoln("Deleting Authz Policy ...")
-	if !c.dryrun {
-		if res := c.configStoreCache.Get(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), obj.Name, obj.Namespace); res != nil {
-			err := c.configStoreCache.Delete(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), obj.Name, obj.Namespace)
-			if err != nil {
-				return fmt.Errorf("error deleting authz policy %v, error: %v", obj.Name, err.Error())
-			}
-		} else {
-			log.Infoln("no authorization policy resource found in cache, authorization policy name:", obj.Name)
-			return nil
-		}
-	} else {
-		err := c.findDeleteDryrunResource(obj.Name, obj.Namespace)
-		if err != nil {
-			return fmt.Errorf("error deleting local authz policy file: %v", err.Error())
-		}
-	}
-	return nil
-}
-
-// genAuthzPolicyConfig is a common function uses by multiple events handler, it reads in athenz domain, given authz
-// policy name, namespace and matching service label, derives authz policy config and return it
-func (c *Controller) genAuthzPolicyConfig(signedDomain zms.SignedDomain, apNamespace, apName, svcLabel string) model.Config {
-	domainRBAC := athenz.ConvertAthenzPoliciesIntoRbacModel(signedDomain.Domain, &c.adIndexInformer)
-	return c.rbacProvider.ConvertAthenzModelIntoIstioAuthzPolicy(domainRBAC, apNamespace, apName, svcLabel)
 }
