+++ pkg/istio/rbac/common/util.go
@@ -1,15 +1,21 @@
-// Copyright 2021, Verizon Media Inc.
+// Copyright 2019, Verizon Media Inc.
 // Licensed under the terms of the 3-Clause BSD license. See LICENSE file in
 // github.com/yahoo/k8s-athenz-istio-auth for terms.
 package common
 
 import (
 	"fmt"
+	"github.com/ghodss/yaml"
 	"github.com/gogo/protobuf/proto"
 	"github.com/yahoo/athenz/clients/go/zms"
+	"github.com/yahoo/k8s-athenz-istio-auth/pkg/log"
+	"io/ioutil"
+	"istio.io/istio/pilot/pkg/config/kube/crd"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pkg/config/schema/collection"
+	"istio.io/istio/pkg/config/schema/collections"
 	"net/http"
+	"os"
 	"regexp"
 	"strings"
 	"time"
@@ -21,6 +27,7 @@ const (
 	ServiceRoleKind              = "ServiceRole"
 	AthenzJwtPrefix              = "athenz/"
 	RequestAuthPrincipalProperty = "request.auth.principal"
+	dryRunStoredFilesDirectory   = "/root/authzpolicy/"
 )
 
 var supportedMethods = map[string]bool{
@@ -36,6 +43,8 @@ var supportedMethods = map[string]bool{
 	"*":                true,
 }
 
+var resourceRegex = regexp.MustCompile(`\A(?P<domain>.*):svc.(?P<svc>[^:]*)[:]?(?P<path>.*)\z`)
+
 type Item struct {
 	Operation model.Event
 	Resource  model.Config
@@ -45,9 +54,47 @@ type Item struct {
 }
 
 type OnCompleteFunc func(err error, item *Item) error
-type checkAnnotaion func(model.Config) bool
+type additionalCheck func(model.Config) bool
 
-var resourceRegex = regexp.MustCompile(`\A(?P<domain>.*):svc.(?P<svc>[^:]*)[:]?(?P<path>.*)\z`)
+type EventHandler interface {
+	Add(item *Item) error
+	Update(item *Item) error
+	Delete(item *Item) error
+}
+
+type DryRunHandler struct{}
+
+func (d *DryRunHandler) Add(item *Item) error {
+	return d.createDryrunResource(item)
+}
+
+func (d *DryRunHandler) Update(item *Item) error {
+	return d.createDryrunResource(item)
+}
+
+func (d *DryRunHandler) Delete(item *Item) error {
+	return d.findDeleteDryrunResource(item)
+}
+
+type ApiHandler struct {
+	ConfigStoreCache model.ConfigStoreCache
+}
+
+func (a *ApiHandler) Add(item *Item) error {
+	_, err := a.ConfigStoreCache.Create(item.Resource)
+	return err
+}
+
+func (a *ApiHandler) Update(item *Item) error {
+	_, err := a.ConfigStoreCache.Update(item.Resource)
+	return err
+}
+
+func (a *ApiHandler) Delete(item *Item) error {
+	res := item.Resource
+	err := a.ConfigStoreCache.Delete(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), res.Name, res.Namespace)
+	return err
+}
 
 // MemberToSpiffe parses the Athenz role member into a SPIFFE compliant name.
 // Example: example.domain/sa/service
@@ -244,7 +291,7 @@ func Equal(c1, c2 model.Config) bool {
 // ComputeChangeList checks if two set of config models have any differences, and return its changeList
 // Controller which calls this function is required to pass its own callback handler
 // checkFn is optional, can be set to nil if nothing needs to be checked
-func ComputeChangeList(currentCRs []model.Config, desiredCRs []model.Config, cbHandler OnCompleteFunc, checkFn checkAnnotaion) []*Item {
+func ComputeChangeList(currentCRs []model.Config, desiredCRs []model.Config, cbHandler OnCompleteFunc, checkFn additionalCheck) []*Item {
 	currMap := ConvertSliceToKeyedMap(currentCRs)
 	desiredMap := ConvertSliceToKeyedMap(desiredCRs)
 
@@ -302,4 +349,74 @@ func ComputeChangeList(currentCRs []model.Config, desiredCRs []model.Config, cbH
 	}
 
 	return changeList
+}
+
+// createDryrunResource creates the yaml file of given authorization policy spec in a local directory
+func (d *DryRunHandler) createDryrunResource(item *Item) error {
+	convertedCR := item.Resource
+	authzPolicyName := item.Resource.ConfigMeta.Name
+	namespace := item.Resource.ConfigMeta.Namespace
+	convertedObj, err := crd.ConvertConfig(collections.IstioSecurityV1Beta1Authorizationpolicies, convertedCR)
+	if err != nil {
+		return fmt.Errorf("unable to convert authorization policy config to istio objects, resource name: %v", convertedCR.Name)
+	}
+	configInBytes, err := yaml.Marshal(convertedObj)
+	if err != nil {
+		return fmt.Errorf("could not marshal %v: %v", convertedCR.Name, err)
+	}
+	yamlFileName := authzPolicyName + "--" + namespace + ".yaml"
+	return ioutil.WriteFile(dryRunStoredFilesDirectory+yamlFileName, configInBytes, 0644)
+}
+
+// findDeleteDryrunResource retrieves the yaml file from local directory and deletes it
+func (d *DryRunHandler) findDeleteDryrunResource(item *Item) error {
+	authzPolicyName := item.Resource.ConfigMeta.Name
+	namespace := item.Resource.ConfigMeta.Namespace
+	yamlFileName := authzPolicyName + "--" + namespace + ".yaml"
+	if _, err := os.Stat(dryRunStoredFilesDirectory + yamlFileName); os.IsNotExist(err) {
+		log.Infof("file %s does not exist in local directory", dryRunStoredFilesDirectory+yamlFileName)
+		return nil
+	} else if err != nil {
+		log.Infof("error stating file %s in local directory: %s, error: %s", dryRunStoredFilesDirectory+yamlFileName, err)
+	}
+	log.Infof("deleting file under path: %s\n", dryRunStoredFilesDirectory+yamlFileName)
+	return os.Remove(dryRunStoredFilesDirectory + yamlFileName)
+}
+
+// ReadConvertToModelConfig reads in the authorization policy yaml object and converts it into a model.Config struct
+func ReadConvertToModelConfig(serviceName, namespace string) (*model.Config, error) {
+	// define istio object interface to unmarshall yaml object into
+	item := &crd.IstioKind{Spec: map[string]interface{}{}}
+	yamlFileName := serviceName + "--" + namespace + ".yaml"
+	yamlFile, err := ioutil.ReadFile(dryRunStoredFilesDirectory + yamlFileName)
+	if err != nil {
+		return nil, fmt.Errorf("unable to read yaml file to local directory, err: %s", err)
+	}
+	err = yaml.Unmarshal(yamlFile, item)
+	if err != nil {
+		return nil, fmt.Errorf("unable to unmarshall yaml file, err: %s", err)
+	}
+	config, err := crd.ConvertObject(collections.IstioSecurityV1Beta1Authorizationpolicies, item, "")
+	if err != nil {
+		return nil, fmt.Errorf("unable to convert yaml converted istio object to authorization policy model config, err: %s", err)
+	}
+	return config, nil
+}
+
+// FetchServices walks through the files under a directory, based on the naming convention, it parses the service name
+// from the file name.
+func FetchServices(namespace string) []string {
+	files, err := ioutil.ReadDir(dryRunStoredFilesDirectory)
+	if err != nil {
+		log.Errorf("unable to read yaml files to local directory, err: %s", err)
+	}
+
+	var svcList []string
+	for _, file := range files {
+		if strings.Contains(file.Name(), namespace) {
+			service := strings.Replace(file.Name(), "--"+namespace+".yaml", "", -1)
+			svcList = append(svcList, service)
+		}
+	}
+	return svcList
 }
