+++ pkg/istio/rbac/common/util_test.go
@@ -1,7 +1,11 @@
+// Copyright 2019, Verizon Media Inc.
+// Licensed under the terms of the 3-Clause BSD license. See LICENSE file in
+// github.com/yahoo/k8s-athenz-istio-auth for terms.
 package common
 
 import (
 	"fmt"
+	"github.com/ardielle/ardielle-go/rdl"
 	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 	"github.com/yahoo/athenz/clients/go/zms"
@@ -10,6 +14,11 @@ import (
 	"istio.io/istio/pkg/config/schema/collection"
 	"istio.io/istio/pkg/config/schema/collections"
 	"testing"
+	"time"
+)
+
+var (
+	isNotSystemDisabled, isSystemDisabled int32 = 0, 1
 )
 
 func TestPrincipalToSPIFFE(t *testing.T) {
@@ -585,6 +594,332 @@ func TestConvertSliceToKeyedMap(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			actual := ConvertSliceToKeyedMap(tt.in)
 			assert.Equal(t, tt.expected, actual, "returned map should match the expected map")
+		})
+	}
+}
+
+func TestCheckAthenzMemberExpiry(t *testing.T) {
+	tests := []struct {
+		athenzMember zms.RoleMember
+		expectedRes  bool
+		expectedErr  error
+	}{
+		{
+			athenzMember: zms.RoleMember{
+				MemberName: "user.expired",
+				Expiration: &rdl.Timestamp{
+					Time: time.Now().Add(-time.Hour),
+				},
+			},
+			expectedRes: false,
+			expectedErr: fmt.Errorf("member user.expired is expired"),
+		},
+		{
+			athenzMember: zms.RoleMember{
+				MemberName: "user.valid",
+				Expiration: &rdl.Timestamp{
+					Time: time.Now().Add(time.Hour),
+				},
+			},
+			expectedRes: true,
+			expectedErr: nil,
+		},
+	}
+
+	for _, tt := range tests {
+		actual, err := CheckAthenzMemberExpiry(&tt.athenzMember)
+		assert.Equal(t, tt.expectedRes, actual, "check member expiry result should be equal to expected result")
+		assert.Equal(t, tt.expectedErr, err, "error should be equal")
+	}
+}
+
+func TestCheckAthenzSystemDisabled(t *testing.T) {
+	tests := []struct {
+		athenzMember zms.RoleMember
+		expectedRes  bool
+		expectedErr  error
+	}{
+		{
+			athenzMember: zms.RoleMember{
+				MemberName:     "user.expired",
+				SystemDisabled: &isSystemDisabled,
+			},
+			expectedRes: false,
+			expectedErr: fmt.Errorf("member user.expired is system disabled"),
+		},
+		{
+			athenzMember: zms.RoleMember{
+				MemberName:     "user.valid",
+				SystemDisabled: &isNotSystemDisabled,
+			},
+			expectedRes: true,
+			expectedErr: nil,
+		},
+	}
+
+	for _, tt := range tests {
+		actual, err := CheckAthenzSystemDisabled(&tt.athenzMember)
+		assert.Equal(t, tt.expectedRes, actual, "check member system disabled result should be equal to expected result")
+		assert.Equal(t, tt.expectedErr, err, "error should be equal")
+	}
+}
+
+func TestComputeChangeList(t *testing.T) {
+
+	cbHandler := func(err error, item *Item) error {
+		return err
+	}
+
+	type input struct {
+		current   []model.Config
+		desired   []model.Config
+		cbHandler OnCompleteFunc
+	}
+	tests := []struct {
+		name           string
+		input          input
+		expectedOutput []Item
+	}{
+		{
+			name:           "should return empty change list for empty current and desired list",
+			input:          input{},
+			expectedOutput: make([]Item, 0),
+		},
+		{
+			name: "should add create operations for new items on the desired list",
+			input: input{
+				current: []model.Config{},
+				desired: []model.Config{
+					newSr("test-ns", "svc-role"),
+					newSrb("test-ns", "svc-role"),
+				},
+				cbHandler: cbHandler,
+			},
+			expectedOutput: []Item{
+				{
+					Operation:       model.EventAdd,
+					Resource:        newSr("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventAdd,
+					Resource:        newSrb("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+			},
+		},
+		{
+			name: "should add update operations for changed items on the desired list",
+			input: input{
+				current: []model.Config{
+					newSr("test-ns", "svc-role"),
+					newSrb("test-ns", "svc-role"),
+					newSr("another-ns", "backend-writer"),
+					newSrb("another-ns", "backend-writer"),
+				},
+				desired: []model.Config{
+					updatedSr("test-ns", "svc-role"),
+					updatedSrb("test-ns", "svc-role"),
+					newSr("another-ns", "backend-writer"),
+					newSrb("another-ns", "backend-writer"),
+				},
+				cbHandler: cbHandler,
+			},
+			expectedOutput: []Item{
+				{
+					Operation:       model.EventUpdate,
+					Resource:        updatedSr("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventUpdate,
+					Resource:        updatedSrb("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+			},
+		},
+		{
+			name: "should add delete operation for deleted items on the desired list",
+			input: input{
+				current: []model.Config{
+					newSr("test-ns", "svc-role"),
+					newSrb("test-ns", "svc-role"),
+				},
+				desired:   []model.Config{},
+				cbHandler: cbHandler,
+			},
+			expectedOutput: []Item{
+				{
+					Operation:       model.EventDelete,
+					Resource:        newSr("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventDelete,
+					Resource:        newSrb("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+			},
+		},
+		{
+			name: "should add create,update and delete operations based on the desired list",
+			input: input{
+				current: []model.Config{
+					newSr("test-ns", "svc-role"),
+					newSrb("test-ns", "svc-role"),
+					newSr("some-ns", "frontend-reader"),
+					newSrb("some-ns", "frontend-reader"),
+				},
+				desired: []model.Config{
+					updatedSr("test-ns", "svc-role"),
+					updatedSrb("test-ns", "svc-role"),
+					newSr("another-ns", "backend-writer"),
+					newSrb("another-ns", "backend-writer"),
+				},
+				cbHandler: cbHandler,
+			},
+			expectedOutput: []Item{
+				{
+					Operation:       model.EventUpdate,
+					Resource:        updatedSr("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventUpdate,
+					Resource:        updatedSrb("test-ns", "svc-role"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventAdd,
+					Resource:        updatedSr("another-ns", "backend-writer"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventAdd,
+					Resource:        updatedSrb("another-ns", "backend-writer"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventDelete,
+					Resource:        newSr("some-ns", "frontend-reader"),
+					CallbackHandler: cbHandler,
+				},
+				{
+					Operation:       model.EventDelete,
+					Resource:        newSrb("some-ns", "frontend-reader"),
+					CallbackHandler: cbHandler,
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			actualChangeList := ComputeChangeList(tt.input.current, tt.input.desired, tt.input.cbHandler, nil)
+			assert.Equal(t, len(tt.expectedOutput), len(actualChangeList), "len(expectedChangeList) and len(actualChangeList) should match")
+			for i, expectedItem := range tt.expectedOutput {
+				assert.Equal(t, expectedItem.Operation, actualChangeList[i].Operation, fmt.Sprintf("operation on changelist[%d] does not match with expected", i))
+				assert.Equal(t, expectedItem.Operation, actualChangeList[i].Operation, fmt.Sprintf("operation on changelist[%d] does not match with expected", i))
+			}
+		})
+	}
+}
+
+func updatedSr(ns, role string) model.Config {
+	srSpec := &v1alpha1.ServiceRole{
+		Rules: []*v1alpha1.AccessRule{
+			{
+				Services: []string{WildCardAll},
+				Methods:  []string{"GET"},
+				Constraints: []*v1alpha1.AccessRule_Constraint{
+					{
+						Key:    ConstraintSvcKey,
+						Values: []string{"test-svc"},
+					},
+				},
+			},
+			{
+				Services: []string{WildCardAll},
+				Methods:  []string{"POST"},
+				Constraints: []*v1alpha1.AccessRule_Constraint{
+					{
+						Key:    ConstraintSvcKey,
+						Values: []string{"test-svc"},
+					},
+				},
+			},
+		},
+	}
+	return NewConfig(collections.IstioRbacV1Alpha1Serviceroles, ns, role, srSpec)
+}
+
+func updatedSrb(ns, role string) model.Config {
+	srbSpec := &v1alpha1.ServiceRoleBinding{
+		RoleRef: &v1alpha1.RoleRef{
+			Kind: ServiceRoleKind,
+			Name: role,
+		},
+		Subjects: []*v1alpha1.Subject{
+			{
+				User: "test-user",
+			},
+			{
+				User: "another.client.user",
+			},
+		},
+	}
+	return NewConfig(collections.IstioRbacV1Alpha1Servicerolebindings, ns, role, srbSpec)
+}
+
+func TestEqual(t *testing.T) {
+	tests := []struct {
+		name     string
+		in1      model.Config
+		in2      model.Config
+		expected bool
+	}{
+		{
+			name:     "should return true for empty model.Config items",
+			in1:      model.Config{},
+			in2:      model.Config{},
+			expected: true,
+		},
+		{
+			name:     "should return false for different model.Config item names but same spec",
+			in1:      newSr("test-ns", "my-role"),
+			in2:      newSr("test-ns", "another-role"),
+			expected: false,
+		},
+		{
+			name:     "should return false for different model.Config item names and different spec",
+			in1:      newSr("test-ns", "my-role"),
+			in2:      updatedSr("test-ns", "another-role"),
+			expected: false,
+		},
+		{
+			name:     "should return false for same model.Config item names but different spec",
+			in1:      newSr("test-ns", "my-role"),
+			in2:      updatedSr("test-ns", "my-role"),
+			expected: false,
+		},
+		{
+			name:     "should return false for different model.Config item types but same names",
+			in1:      newSr("test-ns", "my-role"),
+			in2:      newSrb("test-ns", "my-role"),
+			expected: false,
+		},
+		{
+			name:     "should return true for same model.Config item names and spec",
+			in1:      newSr("test-ns", "my-role"),
+			in2:      newSr("test-ns", "my-role"),
+			expected: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			actual := Equal(tt.in1, tt.in2)
+			assert.Equal(t, tt.expected, actual, "comparison result should be equal to expected")
 		})
 	}
 }
