+++ pkg/istio/rbac/v1/provider.go
@@ -4,17 +4,12 @@
 package v1
 
 import (
-	"github.com/yahoo/athenz/clients/go/zms"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/athenz"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/rbac/common"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/log"
-	"istio.io/api/security/v1beta1"
-	workloadv1beta1 "istio.io/api/type/v1beta1"
-	"istio.io/istio/pkg/config/schema/collections"
-	"regexp"
-
 	"istio.io/istio/pilot/pkg/model"
+	"istio.io/istio/pkg/config/schema/collections"
 	"istio.io/istio/pkg/config/validation"
 )
 
@@ -32,7 +27,7 @@ func NewProvider(enableOriginJwtSubject bool) rbac.Provider {
 // ConvertAthenzModelIntoIstioRbac converts the Athenz RBAC model into the list of Istio Authorization V1 specific
 // RBAC custom resources (ServiceRoles, ServiceRoleBindings)
 // The idea is that with a given input model, the function should always return the same output list of resources
-func (p *v1) ConvertAthenzModelIntoIstioRbac(m athenz.Model) []model.Config {
+func (p *v1) ConvertAthenzModelIntoIstioRbac(m athenz.Model, serviceName string, svcLabel string) []model.Config {
 
 	out := make([]model.Config, 0)
 
@@ -99,7 +94,7 @@ func (p *v1) ConvertAthenzModelIntoIstioRbac(m athenz.Model) []model.Config {
 }
 
 // GetCurrentIstioRbac returns the ServiceRole and ServiceRoleBinding resources for the specified model's namespace
-func (p *v1) GetCurrentIstioRbac(m athenz.Model, csc model.ConfigStoreCache) []model.Config {
+func (p *v1) GetCurrentIstioRbac(m athenz.Model, csc model.ConfigStoreCache, serviceName string, dryrun bool) []model.Config {
 
 	sr, err := csc.List(collections.IstioRbacV1Alpha1Serviceroles.Resource().GroupVersionKind(), m.Namespace)
 	if err != nil {
@@ -112,156 +107,4 @@ func (p *v1) GetCurrentIstioRbac(m athenz.Model, csc model.ConfigStoreCache) []m
 	}
 
 	return append(sr, srb...)
-}
-
-// ConvertAthenzModelIntoIstioAuthzPolicy converts the Athenz RBAC model into Istio Authorization V1Beta1 specific
-// RBAC custom resource (AuthorizationPolicy)
-func (p *v1) ConvertAthenzModelIntoIstioAuthzPolicy(athenzModel athenz.Model, namespace string, serviceName string, svcLabel string) model.Config {
-	// authz policy is created per service. each rule is created by each role, and form the rules under
-	// this authz policy.
-	var out model.Config
-	// form authorization config meta
-	// namespace: service's namespace
-	// name: service's name
-	schema := collections.IstioSecurityV1Beta1Authorizationpolicies
-	out.ConfigMeta = model.ConfigMeta{
-		Type:      schema.Resource().Kind(),
-		Group:     schema.Resource().Group(),
-		Version:   schema.Resource().Version(),
-		Namespace: namespace,
-		Name:      serviceName,
-	}
-
-	// matching label, same with the service label
-	spec := &v1beta1.AuthorizationPolicy{}
-
-	spec.Selector = &workloadv1beta1.WorkloadSelector{
-		MatchLabels: map[string]string{"svc": svcLabel},
-	}
-
-	// generating rules, iterate through assertions, find the one match with desired format.
-	var rules []*v1beta1.Rule
-	for role, assertions := range athenzModel.Rules {
-		for _, assert := range assertions {
-			// assert.Resource contains the svc information that needs to parse and match
-			svc, path, err := common.ParseAssertionResource(zms.DomainName(athenz.NamespaceToDomain(namespace)), assert)
-			if err != nil {
-				continue
-			}
-			// if svc match with current svc, process it and add it to the rules
-			// note that svc defined on athenz can be a regex, need to match the pattern
-			res, e := regexp.MatchString(svc, svcLabel)
-			if e != nil {
-				log.Errorln("error matching string: ", e.Error())
-				continue
-			}
-			if !res {
-				log.Errorf("athenz svc %s does not match with current svc %s", svc, svcLabel)
-				continue
-			}
-			rule := &v1beta1.Rule{}
-			// form rule.From, must initialize internal source here
-			from_principal := &v1beta1.Rule_From{
-				Source: &v1beta1.Source{},
-			}
-			from_requestPrincipal := &v1beta1.Rule_From{
-				Source: &v1beta1.Source{},
-			}
-			// role name should match zms resource name
-			for _, roleName := range athenzModel.Roles {
-				if roleName == role {
-					// add function to enableOriginJwtSubject, following code assume enableOriginJwtSubject is true by default
-					for _, roleMember := range athenzModel.Members[roleName] {
-						res, err := common.CheckAthenzMemberExpiry(roleMember)
-						if err != nil {
-							log.Errorf("error when checking athenz member expiration date, skipping current member: %s, error: %s", roleMember.MemberName, err)
-							continue
-						}
-						if !res {
-							log.Infoln("member expired, skip adding member to authz policy resource, member: ", roleMember.MemberName)
-							continue
-						}
-						res, err = common.CheckAthenzSystemDisabled(roleMember)
-						if err != nil {
-							log.Errorf("error when checking athenz member system disabled, skipping current member: %s, error: %s", roleMember.MemberName, err)
-							continue
-						}
-						if !res {
-							log.Infoln("member expired, skip adding member to authz policy resource, member: ", roleMember.MemberName)
-							continue
-						}
-
-						spiffeName, err := common.MemberToSpiffe(roleMember)
-						if err != nil {
-							log.Errorln("error converting role name to spiffeName: ", err.Error())
-							continue
-						}
-						from_principal.Source.Principals = append(from_principal.Source.Principals, spiffeName)
-						if p.enableOriginJwtSubject {
-							originJwtName, err := common.MemberToOriginJwtSubject(roleMember)
-							if err != nil {
-								log.Errorln(err.Error())
-								continue
-							}
-							from_requestPrincipal.Source.RequestPrincipals = append(from_requestPrincipal.Source.RequestPrincipals, originJwtName)
-						}
-					}
-					//add role spiffee for role certificate
-					roleSpiffeName, err := common.RoleToSpiffe(string(athenzModel.Name), string(roleName))
-					if err != nil {
-						log.Println("error when convert role to spiffe name: ", err.Error())
-						continue
-					}
-					from_principal.Source.Principals = append(from_principal.Source.Principals, roleSpiffeName)
-				}
-			}
-			rule.From = append(rule.From, from_principal)
-			rule.From = append(rule.From, from_requestPrincipal)
-			// form rules_to
-			rule_to := &v1beta1.Rule{}
-			_, err = common.ParseAssertionEffect(assert)
-			if err != nil {
-				log.Debugf(err.Error())
-				continue
-			}
-			method, err := common.ParseAssertionAction(assert)
-			if err != nil {
-				log.Debugf(err.Error())
-				continue
-			}
-			// form rule.To
-			to := &v1beta1.Rule_To{
-				Operation: &v1beta1.Operation{
-					Methods: []string{method},
-				},
-			}
-			if path != "" {
-				to.Operation.Paths = []string{path}
-			}
-			rule_to.To = append(rule_to.To, to)
-			rules = append(rules, rule_to)
-			rules = append(rules, rule)
-		}
-	}
-	spec.Rules = rules
-	out.Spec = spec
-	return out
-}
-
-// GetCurrentIstioRbac returns the authorization policies resources for the specified model's namespace
-// if serviceName is "", return the all the authorization policies in the given namespace,
-// if serviceName is specific, return single authorization policy matching with serviceName.
-func (p *v1) GetCurrentIstioAuthzPolicy(m athenz.Model, csc model.ConfigStoreCache, serviceName string) []model.Config {
-	if serviceName == "" {
-		apList, err := csc.List(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), m.Namespace)
-		if err != nil {
-			log.Errorf("Error listing the Authorization Policy resources in the namespace: %s", m.Namespace)
-		}
-		return apList
-	}
-	ap := csc.Get(collections.IstioSecurityV1Beta1Authorizationpolicies.Resource().GroupVersionKind(), serviceName, m.Namespace)
-	if ap != nil {
-		return []model.Config{*ap}
-	}
-	return []model.Config{}
 }
